added validation for post (user)

+++ controllers/feed.js
@@ -1,7 +1,10 @@
-const { validationResult } = require("express-validator/check");
-const Post = require("../models/post");
 const fs = require("fs");
 const path = require("path");
+
+const { validationResult } = require("express-validator/check");
+
+const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -17,13 +20,11 @@ exports.getPosts = (req, res, next) => {
         .limit(perPage);
     })
     .then(posts => {
-      res
-        .status(200)
-        .json({
-          message: "Fetched post successfully!",
-          posts: posts,
-          totalItems: totalItems
-        });
+      res.status(200).json({
+        message: "Fetched post successfully!",
+        posts: posts,
+        totalItems: totalItems
+      });
     })
     .catch(err => {
       if (!err.statusCode) {
@@ -51,19 +52,29 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Nikhil" }
+    creator: req.userId
   });
+
   post
     .save()
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: "Post created Succesfully!",
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -116,6 +127,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -145,11 +163,22 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
       //check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.post.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Deleted Post" });
     })
     .catch(err => {
